+++ python/metatomic_torch/build-backend/backend.py
@@ -27,7 +27,8 @@ build_sdist = build_meta.build_sdist
 # Special dependencies to build the wheels
 def get_requires_for_build_wheel(config_settings=None):
     defaults = build_meta.get_requires_for_build_wheel(config_settings)
-    return defaults + [
-        TORCH_DEP,
-        "metatensor-torch >=0.8.0,<0.9.0",
-    ]
+    return defaults + [TORCH_DEP]
+
+
+def build_editable(wheel_directory, config_settings=None, metadata_directory=None):
+    raise RuntimeError("metatomic-torch does not support editable installation yet")