+++ python/metatomic_torch/metatomic/torch/ase_calculator.py
@@ -18,6 +18,7 @@ from . import (
     ModelOutput,
     System,
     load_atomistic_model,
+    pick_device,
     register_autograd_neighbors,
 )
 
@@ -66,6 +67,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         extensions_directory=None,
         check_consistency=False,
         device=None,
+        variants: Optional[Dict[str, Optional[str]]] = None,
         non_conservative=False,
         do_gradients_with_energy=True,
         uncertainty_threshold=0.1,
@@ -88,6 +90,12 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             running, defaults to False.
         :param device: torch device to use for the calculation. If ``None``, we will try
             the options in the model's ``supported_device`` in order.
+        :param variants: dictionary mapping output names to a variant that should be
+            used for the calculations (e.g. ``{"energy": "PBE"}``). If ``"energy"`` is
+            set to a variant also the uncertainty and non conservative outputs will be
+            taken from this variant. This behaviour can be overriden by setting the
+            corresponding keys explicitly to ``None`` or to another value (e.g.
+            ``{"energy_uncertainty": "r2scan"}``).
         :param non_conservative: if ``True``, the model will be asked to compute
             non-conservative forces and stresses. This can afford a speed-up,
             potentially at the expense of physical correctness (especially in molecular
@@ -111,6 +119,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         self.parameters = {
             "extensions_directory": extensions_directory,
             "check_consistency": bool(check_consistency),
+            "variants": variants,
             "non_conservative": bool(non_conservative),
             "do_gradients_with_energy": bool(do_gradients_with_energy),
             "additional_outputs": additional_outputs,
@@ -142,35 +151,12 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             raise TypeError(f"unknown type for model: {type(model)}")
 
         self.parameters["device"] = str(device) if device is not None else None
-        # check if the model supports the requested device
+        # get the best device according what the model supports and what's available on
+        # the current machine
         capabilities = model.capabilities()
-        if device is None:
-            device = _find_best_device(capabilities.supported_devices)
-        else:
-            device = torch.device(device)
-            device_is_supported = False
-
-            for supported in capabilities.supported_devices:
-                try:
-                    supported = torch.device(supported)
-                except RuntimeError as e:
-                    warnings.warn(
-                        "the model contains an invalid device in `supported_devices`: "
-                        f"{e}",
-                        stacklevel=2,
-                    )
-                    continue
-
-                if supported.type == device.type:
-                    device_is_supported = True
-                    break
-
-            if not device_is_supported:
-                raise ValueError(
-                    f"This model does not support the requested device ({device}), "
-                    "the following devices are supported: "
-                    f"{capabilities.supported_devices}"
-                )
+        self._device = torch.device(
+            pick_device(capabilities.supported_devices, self.parameters["device"])
+        )
 
         if capabilities.dtype in STR_TO_DTYPE:
             self._dtype = STR_TO_DTYPE[capabilities.dtype]
@@ -178,6 +164,51 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             raise ValueError(
                 f"found unexpected dtype in model capabilities: {capabilities.dtype}"
             )
+
+        self._energy_key = "energy"
+        self._energy_uq_key = "energy_uncertainty"
+        self._nc_forces_key = "non_conservative_forces"
+        self._nc_stress_key = "non_conservative_stress"
+
+        if variants:
+            if "energy" in variants:
+                self._energy_key += f"/{variants['energy']}"
+                self._energy_uq_key += f"/{variants['energy']}"
+                self._nc_forces_key += f"/{variants['energy']}"
+                self._nc_stress_key += f"/{variants['energy']}"
+
+            if "energy_uncertainty" in variants:
+                if variants["energy_uncertainty"] is None:
+                    self._energy_uq_key = "energy_uncertainty"
+                else:
+                    self._energy_uq_key += f"/{variants['energy_uncertainty']}"
+
+            if non_conservative:
+                if (
+                    "non_conservative_stress" in variants
+                    and "non_conservative_forces" in variants
+                    and (
+                        (variants["non_conservative_stress"] is None)
+                        != (variants["non_conservative_forces"] is None)
+                    )
+                ):
+                    raise ValueError(
+                        "if both 'non_conservative_stress' and "
+                        "'non_conservative_forces' are present in `variants`, they "
+                        "must either be both `None` or both not `None`."
+                    )
+
+                if "non_conservative_forces" in variants:
+                    if variants["non_conservative_forces"] is None:
+                        self._nc_forces_key = "non_conservative_forces"
+                    else:
+                        self._nc_forces_key += f"/{variants['non_conservative_forces']}"
+
+                if "non_conservative_stress" in variants:
+                    if variants["non_conservative_stress"] is None:
+                        self._nc_stress_key = "non_conservative_stress"
+                    else:
+                        self._nc_stress_key += f"/{variants['non_conservative_stress']}"
 
         if additional_outputs is None:
             self._additional_output_requests = {}
@@ -192,20 +223,22 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
             self._additional_output_requests = additional_outputs
 
-        self._device = device
         self._model = model.to(device=self._device)
 
         self._calculate_uncertainty = (
-            "energy_uncertainty" in self._model.capabilities().outputs
+            self._energy_uq_key in self._model.capabilities().outputs
             # we require per-atom uncertainties to capture local effects
-            and self._model.capabilities().outputs["energy_uncertainty"].per_atom
+            and self._model.capabilities().outputs[self._energy_uq_key].per_atom
             and uncertainty_threshold is not None
         )
 
         if self._calculate_uncertainty:
             assert uncertainty_threshold is not None
             if uncertainty_threshold <= 0.0:
-                raise ValueError("`uncertainty_threshold` can not be negative")
+                raise ValueError(
+                    f"`uncertainty_threshold` is {uncertainty_threshold} but must "
+                    "be positive"
+                )
 
         # We do our own check to verify if a property is implemented in `calculate()`,
         # so we pretend to be able to compute all properties ASE knows about.
@@ -368,7 +401,12 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             )
             outputs.update(self._additional_output_requests)
             if calculate_energy and self._calculate_uncertainty:
-                outputs["energy_uncertainty"] = _get_energy_uncertainty_output()
+                outputs[self._energy_uq_key] = ModelOutput(
+                    quantity="energy",
+                    unit="eV",
+                    per_atom=True,
+                    explicit_gradients=[],
+                )
 
             capabilities = self._model.capabilities()
             for name in outputs.keys():
@@ -426,10 +464,10 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             run_options,
             check_consistency=self.parameters["check_consistency"],
         )
-        energy = outputs["energy"]
+        energy = outputs[self._energy_key]
 
         with record_function("MetatomicCalculator::sum_energies"):
-            if run_options.outputs["energy"].per_atom:
+            if run_options.outputs[self._energy_key].per_atom:
                 assert len(energy) == 1
                 assert energy.sample_names == ["system", "atom"]
                 assert torch.all(energy.block().samples["system"] == 0)
@@ -445,17 +483,16 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
         with record_function("ASECalculator::uncertainty_warning"):
             if calculate_energy and self._calculate_uncertainty:
-                uncertainty = outputs["energy_uncertainty"].block().values
+                uncertainty = outputs[self._energy_uq_key].block().values
                 assert uncertainty.shape == (len(atoms), 1)
                 uncertainty = uncertainty.detach().cpu().numpy()
 
                 threshold = self.parameters["uncertainty_threshold"]
                 if np.any(uncertainty > threshold):
                     warnings.warn(
-                        "Some of the atomic energy uncertainties are larger than"
-                        f"the threshold of {threshold} eV."
-                        "The prediction is above the threshold for atoms"
-                        f" {np.where(uncertainty > threshold)[0]}.",
+                        "Some of the atomic energy uncertainties are larger than the "
+                        f"threshold of {threshold} eV. The prediction is above the "
+                        f"threshold for atoms {np.where(uncertainty > threshold)[0]}.",
                         stacklevel=2,
                     )
 
@@ -497,9 +534,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
             if calculate_forces:
                 if self.parameters["non_conservative"]:
-                    forces_values = (
-                        outputs["non_conservative_forces"].block().values.detach()
-                    )
+                    forces_values = outputs[self._nc_forces_key].block().values.detach()
                     # remove any spurious net force
                     forces_values = forces_values - forces_values.mean(
                         dim=0, keepdim=True
@@ -512,9 +547,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
 
             if calculate_stress:
                 if self.parameters["non_conservative"]:
-                    stress_values = (
-                        outputs["non_conservative_stress"].block().values.detach()
-                    )
+                    stress_values = outputs[self._nc_stress_key].block().values.detach()
                 else:
                     stress_values = strain.grad / atoms.cell.volume
                 stress_values = stress_values.reshape(3, 3)
@@ -603,7 +636,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
             options=options,
             check_consistency=self.parameters["check_consistency"],
         )
-        energies = predictions["energy"]
+        energies = predictions[self._energy_key]
 
         results_as_numpy_arrays = {
             "energy": energies.block().values.detach().cpu().numpy().flatten().tolist()
@@ -611,7 +644,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         if compute_forces_and_stresses:
             if self.parameters["non_conservative"]:
                 results_as_numpy_arrays["forces"] = (
-                    predictions["non_conservative_forces"]
+                    predictions[self._nc_forces_key]
                     .block()
                     .values.squeeze(-1)
                     .detach()
@@ -635,7 +668,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 if all(atoms.pbc.all() for atoms in atoms_list):
                     results_as_numpy_arrays["stress"] = [
                         s
-                        for s in predictions["non_conservative_stress"]
+                        for s in predictions[self._nc_stress_key]
                         .block()
                         .values.squeeze(-1)
                         .detach()
@@ -651,7 +684,7 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 if all(atoms.pbc.all() for atoms in atoms_list):
                     results_as_numpy_arrays["stress"] = [
                         strain.grad.cpu().numpy() / atoms.cell.volume
-                        for strain, atoms in zip(strains, atoms_list)
+                        for strain, atoms in zip(strains, atoms_list, strict=False)
                     ]
         if was_single:
             for key, value in results_as_numpy_arrays.items():
@@ -687,16 +720,16 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
         else:
             output.per_atom = False
 
-        metatensor_outputs = {"energy": output}
+        metatensor_outputs = {self._energy_key: output}
         if calculate_forces and self.parameters["non_conservative"]:
-            metatensor_outputs["non_conservative_forces"] = ModelOutput(
+            metatensor_outputs[self._nc_forces_key] = ModelOutput(
                 quantity="force",
                 unit="eV/Angstrom",
                 per_atom=True,
             )
 
         if calculate_stress and self.parameters["non_conservative"]:
-            metatensor_outputs["non_conservative_stress"] = ModelOutput(
+            metatensor_outputs[self._nc_stress_key] = ModelOutput(
                 quantity="pressure",
                 unit="eV/Angstrom^3",
                 per_atom=False,
@@ -713,49 +746,6 @@ class MetatomicCalculator(ase.calculators.calculator.Calculator):
                 raise ValueError(f"this model does not support '{key}' output")
 
         return metatensor_outputs
-
-
-def _find_best_device(devices: List[str]) -> torch.device:
-    """
-    Find the best device from the list of ``devices`` that is available to the current
-    PyTorch installation.
-    """
-
-    for device in devices:
-        if device == "cpu":
-            return torch.device("cpu")
-        elif device == "cuda":
-            if torch.cuda.is_available():
-                return torch.device("cuda")
-            else:
-                LOGGER.warning(
-                    "the model suggested to use CUDA devices before CPU, "
-                    "but we are unable to find it"
-                )
-        elif device == "mps":
-            if (
-                hasattr(torch.backends, "mps")
-                and torch.backends.mps.is_built()
-                and torch.backends.mps.is_available()
-            ):
-                return torch.device("mps")
-            else:
-                LOGGER.warning(
-                    "the model suggested to use MPS devices before CPU, "
-                    "but we are unable to find it"
-                )
-        else:
-            warnings.warn(
-                f"unknown device in the model's `supported_devices`: '{device}'",
-                stacklevel=2,
-            )
-
-    warnings.warn(
-        "could not find a valid device in the model's `supported_devices`, "
-        "falling back to CPU",
-        stacklevel=2,
-    )
-    return torch.device("cpu")
 
 
 def _compute_ase_neighbors(atoms, options, dtype, device):
@@ -858,15 +848,6 @@ def _full_3x3_to_voigt_6_stress(stress):
             (stress[0, 2] + stress[2, 0]) / 2.0,
             (stress[0, 1] + stress[1, 0]) / 2.0,
         ]
-    )
-
-
-def _get_energy_uncertainty_output():
-    return ModelOutput(
-        quantity="energy",
-        unit="eV",
-        per_atom=True,
-        explicit_gradients=[],
     )