+++ metatomic-torch/CHANGELOG.md
@@ -17,12 +17,37 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 ### Removed
 -->
 
+### Removed
+
+Dropped support for deprecated Python 3.9, now requires 3.10 as minimum version
+
+## [Version 0.1.5](https://github.com/metatensor/metatomic/releases/tag/metatomic-torch-v0.1.5) - 2025-10-06
+
+### Added
+
+- Two new functions, `metatomic.torch.save_buffer` and
+  `metatomic.torch.load_system_buffer`, allow to serialize and deserialize
+  `System` objects to and from `torch.Tensor`
+
+### Changed
+
+- `metatomic.torch.save` and `metatomic.torch.load_system` are now implemented
+  in C++
+- We now requires at least cmake v3.22 to compile metatomic-torch
+
+## [Version 0.1.4](https://github.com/metatensor/metatomic/releases/tag/metatomic-torch-v0.1.4) - 2025-09-11
+
 ### Added
 
 - The code is now compatible with metatensor-torch v0.8.0
 - The code is now compatible torch v2.8
+
 - `System.to` accepts a `non_blocking` argument, with the same meaning as
   `torch.Tensor.to`.
+- The ASE `MetatomicCalculator` will now send warnings if the model predicts a
+  high per-atom uncertainty for its energy output.
+- We now have two new standard outputs: `positions` and `momenta`, which can be
+  used for direct structure prediction and bypassing time integration.
 
 ## [Version 0.1.3](https://github.com/metatensor/metatomic/releases/tag/metatomic-torch-v0.1.3) - 2025-07-25